extract out tables to shared component
