+++ src/features/Dashboard/components/WalletBreakdown/index.ts
@@ -1,2 +1,3 @@
 export { WalletBreakdown } from './WalletBreakdown';
 export * from './helpers';
+export { BalancesDetailProps } from './types';
