+++ src/features/Dashboard/components/WalletBreakdown/types.ts
@@ -1,26 +1,6 @@
-import { IAccount, Fiat, TAddress, TUuid } from '@types';
+import { Balance, Fiat, IAccount } from '@types';
 
-export interface BalanceAccount {
-  address: TAddress;
-  ticker: string;
-  amount: number;
-  fiatValue: number;
-  label: string;
-}
-
-export interface Balance {
-  id?: string;
-  name: string;
-  amount: number;
-  fiatValue: number;
-  ticker: string;
-  isOther?: boolean;
-  exchangeRate?: number;
-  accounts?: BalanceAccount[];
-  uuid?: TUuid;
-}
-
-export interface WalletBreakdownProps {
+export interface BalancesDetailProps {
   balances: Balance[];
   totalFiatValue: number;
   fiat: Fiat;
