+++ src/types/index.ts
@@ -103,3 +103,4 @@ export { TAction, TStateGetter } from './action';
 export { TxParcel } from '@utils/useTxMulti/types'; // Specific re-export to avoid Circular deps
 export { TURL } from './url';
 export { TBN } from './bigNumber';
+export { Balance, BalanceAccount, BalanceDetailsTableProps } from './balanceDisplays';
